refactoring and improovements
