+++ src/test/java/Orders/OrdersCreationTests.java
@@ -1,3 +1,6 @@
+package Orders;
+
+import Users.UserSteps;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
