+++ src/test/java/Users/EditUserProfileTests.java
@@ -1,10 +1,12 @@
+package Users;
+
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
 
-import static practicum.Constants.USER_EXIST_ERROR;
+import static practicum.Constants.NOT_AUTHORISED_ERROR;
 
-public class UserCreationTests {
+public class EditUserProfileTests {
 
     private UserSteps userSteps = new UserSteps();
 
@@ -14,13 +16,13 @@ public class UserCreationTests {
     }
 
     @Test
-    public void uniqUserCreationTest(){
-        userSteps.verifySuccessfulUserLoginResponse();
+    public void editUserProfileTest(){
+        userSteps.editUserData().verifyEditedProfile();
     }
 
     @Test
-    public void createExistedUser(){
-        userSteps.registerUser().verifyUnsuccessfulResponse(403, USER_EXIST_ERROR);
+    public void notEditUserProfileTest(){
+        userSteps.editUserDataNoAuth().verifyUnsuccessfulResponse(401, NOT_AUTHORISED_ERROR);
     }
 
     @After
