+++ src/test/java/Users/UserLoginTests.java
@@ -1,3 +1,5 @@
+package Users;
+
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
